+++ src/main/java/ladder/domain/LadderResult.java
@@ -1,4 +1,4 @@
-package ladder.model;
+package ladder.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
