+++ src/main/java/ladder/domain/LineUnit.java
@@ -1,4 +1,4 @@
-package ladder.model;
+package ladder.domain;
 
 public class LineUnit {
