+++ src/main/java/ladder/domain/UserName.java
@@ -1,4 +1,4 @@
-package ladder.model;
+package ladder.domain;
 
 import java.util.Objects;
