+++ src/main/java/ladder/domain/VerticalLine.java
@@ -1,4 +1,4 @@
-package ladder.model;
+package ladder.domain;
 
 import java.util.ArrayList;
 import java.util.List;
