+++ src/main/java/ladder/ui/OutputView.java
@@ -1,6 +1,6 @@
 package ladder.ui;
 
-import ladder.model.*;
+import ladder.domain.*;
 
 import java.util.List;
 import java.util.stream.Collectors;
